requirements.txt
Django==3.1.7
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
psycopg2==2.8.6
drf-yasg==1.20.0
